--- Backstage/base/ops.ts
@@ -1,1 +1,0 @@
-export type LogicOps = 'and' | 'or' | 'not'
