+++ Backstage/package.json
@@ -2,7 +2,7 @@
    "name": "cnpr",
    "version": "1.0.0",
    "description": "",
-   "main": "index.js",
+   "main": "src/index.ts",
    "scripts": {
       "test_core_ruleset_assertion": "node test/test_assertions.js",
       "test_compile": "node test/test_compile.js",
@@ -12,8 +12,11 @@
    "author": "",
    "license": "ISC",
    "devDependencies": {
-      "eslint": "^8.20.0",
+      "@typescript-eslint/eslint-plugin": "^5.32.0",
+      "@typescript-eslint/parser": "^5.32.0",
+      "eslint": "^8.21.0",
       "eslint-config-airbnb-base": "^15.0.0",
+      "eslint-import-resolver-typescript": "^3.4.0",
       "eslint-plugin-import": "^2.26.0",
       "json5": "^2.2.1",
       "ts-node": "^10.9.1",
