+++ Backstage/src/core_ruleset/activities/init_skills.js
