+++ Backstage/src/core_ruleset/activity_categories.js
