+++ Backstage/src/core_ruleset/ascension_perks/learn_from_the_past.js
