+++ Backstage/src/core_ruleset/ascension_perks/precautions.js
