+++ Backstage/src/core_ruleset/skill_categories.js
