+++ Backstage/src/core_ruleset/translations/zh_cn/ascension_perks.js
