+++ Backstage/src/core_ruleset/translations/zh_cn/index.js
