+++ Backstage/src/core_ruleset/translations/zh_cn/skill_categories.js
