+++ Backstage/src/core_ruleset/translations/zh_cn/skills.js
