+++ Backstage/src/executor/compute.js
