+++ Backstage/src/ruleset/event.ts
@@ -1,0 +1,14 @@
+import {GameContext} from "../executor/game_context";
+import {Ident} from "../base/uid";
+
+export type EventFunction = (gameContext: GameContext, ...args: any[]) => void
+
+export class Event {
+    readonly ident: Ident
+    readonly event: EventFunction[]
+
+    constructor(ident: Ident, event: EventFunction[]) {
+        this.ident = ident
+        this.event = event
+    }
+}
