+++ Backstage/src/ruleset/index.ts
@@ -1,0 +1,10 @@
+import { Skill } from '../executor/ruleset'
+
+export class CompiledRuleSet {
+   skillCategories: string[]
+   activityCategories: string[]
+
+   events: { [key: string]: Event }
+   modifiers: object // TODO(chuigda): modifier system rework
+   skills: { [key: string]: Skill }
+}
