+++ Backstage/src/ruleset/item_base.ts
@@ -1,0 +1,13 @@
+import { Ident, MaybeTranslationKey } from '../base/uid'
+
+export class ItemBase {
+   readonly ident: Ident
+   readonly name: MaybeTranslationKey
+   readonly description: MaybeTranslationKey
+
+   constructor(ident: Ident, name: MaybeTranslationKey, description: MaybeTranslationKey) {
+      this.ident = ident
+      this.name = name
+      this.description = description
+   }
+}
