+++ Backstage/src/ruleset/ops.ts
@@ -1,0 +1,1 @@
+export type LogicOp = 'and' | 'or' | 'not'
