+++ Backstage/src/util/emergency.js
