+++ Backstage/src/util/stringify.js
