+++ Backstage/src/util/type_assert.js
