+++ Backstage/test/test_assertions.js
@@ -1,5 +1,5 @@
-const coreRuleSet = require('../core_ruleset')
-const { ruleSetAssertion } = require('../assertions')
-const { typeAssert } = require('../util/type_assert')
+const coreRuleSet = require('../src/core_ruleset')
+const { ruleSetAssertion } = require('../src/assertions')
+const { typeAssert } = require('../src/util/type_assert')
 
 typeAssert(coreRuleSet, ruleSetAssertion)
