+++ Backstage/test/test_compile.js
@@ -1,7 +1,7 @@
-const { perfectStringify } = require('../util/stringify')
+const { perfectStringify } = require('../src/util/stringify')
 
-const { emptyRuleSet, compileRuleSet } = require('../executor/ruleset')
-const coreRuleSet = require('../core_ruleset')
+const { emptyRuleSet, compileRuleSet } = require('../src/executor/ruleset')
+const coreRuleSet = require('../src/core_ruleset')
 
 const ruleSet = emptyRuleSet()
 compileRuleSet(ruleSet, coreRuleSet)
