+++ Backstage/tsconfig.json
@@ -1,0 +1,16 @@
+{
+   "compilerOptions": {
+     "module": "commonjs",
+     "noImplicitAny": true,
+     "removeComments": true,
+     "preserveConstEnums": true,
+     "sourceMap": true
+   },
+   "include": [
+      "src/**/*"
+   ],
+   "exclude": [
+      "node_modules",
+      "**/*.spec.ts"
+   ]
+}
